Add unit test on final logits and attention
