+++ python/llm/test/inference_gpu/test_transformers_api_final_logits.py
@@ -14,33 +14,39 @@
 # limitations under the License.
 #
 
-import os
+
+import os, time
 import pytest
+import tempfile
+
 import torch
-from bigdl.llm.transformers import AutoModelForCausalLM
-from transformers import AutoTokenizer
+from bigdl.llm.transformers import AutoModelForCausalLM, AutoModel, AutoModelForSpeechSeq2Seq
+from transformers import LlamaTokenizer, AutoTokenizer
 
 device = os.environ['DEVICE']
 print(f'Running on {device}')
 
-prompt = "Once upon a time, there existed a little girl who liked to have adventures. She wanted to go to places and meet new people, and have fun"
+PROMPT = "Once upon a time, there existed a little girl who liked to have adventures. She wanted to go to places and meet new people, and have fun"
+TEST_MODEL_LIST = [
+    ("MPT-7B", AutoModelForCausalLM, AutoTokenizer, os.environ.get('MPT_7B_ORIGIN_PATH')),
+    ("Falcon-7B", AutoModelForCausalLM, AutoTokenizer, os.environ.get('FALCON_7B_ORIGIN_PATH')),
+]
 
-@pytest.mark.parametrize('Model, Tokenizer, model_path',[
-    (AutoModelForCausalLM, AutoTokenizer, os.environ.get('MPT_7B_ORIGIN_PATH')),
-    # (AutoModelForCausalLM, AutoTokenizer, os.environ.get('LLAMA2_7B_ORIGIN_PATH')),
-    (AutoModelForCausalLM, AutoTokenizer, os.environ.get('FALCON_7B_ORIGIN_PATH')),
-    ])
-def test_optimize_model(Model, Tokenizer, model_path):
+
+@pytest.mark.parametrize('Name, Model, Tokenizer, model_path',TEST_MODEL_LIST)
+def test_optimize_model(Name, Model, Tokenizer, model_path):
     with torch.inference_mode():
         tokenizer = Tokenizer.from_pretrained(model_path, trust_remote_code=True)
-        input_ids = tokenizer.encode(prompt, return_tensors="pt").to(device)
+        input_ids = tokenizer.encode(PROMPT, return_tensors="pt").to(device)
 
         model = Model.from_pretrained(model_path,
                                     load_in_4bit=True,
                                     optimize_model=False,
                                     trust_remote_code=True)
         model = model.to(device)
+
         logits_base_model = (model(input_ids)).logits
+
         model.to('cpu')  # deallocate gpu memory
 
         model = Model.from_pretrained(model_path,
@@ -54,6 +60,11 @@ def test_optimize_model(Model, Tokenizer, model_path):
         tol = 1e-03
         num_false = torch.isclose(logits_optimized_model, logits_base_model, rtol=tol, atol=tol)\
             .flatten().tolist().count(False)
+        
         percent_false = num_false / logits_optimized_model.numel()
+
         assert percent_false < 1e-02
-    
+
+
+if __name__ == '__main__':
+    pytest.main([__file__])
