+++ python/llm/test/run-llm-inference-tests-gpu.sh
@@ -19,7 +19,8 @@ fi
 export OMP_NUM_THREADS=$THREAD_NUM
 pytest ${LLM_INFERENCE_TEST_DIR}/test_transformers_api.py -v -s
 export BIGDL_LLM_XMX_DISABLED=1
-pytest ${LLM_INFERENCE_TEST_DIR}/test_transformers_api_disable_xmx.py -v -s
+pytest ${LLM_INFERENCE_TEST_DIR}/test_transformers_api_final_logits.py -v -s
+pytest ${LLM_INFERENCE_TEST_DIR}/test_transformers_api_attention.py -v -s
 unset BIGDL_LLM_XMX_DISABLED
 
 now=$(date "+%s")
